+++ src/pages/Beta/Airdrop2/ButtonToConnect/index.tsx
