+++ src/pages/Beta/Bridge/QuestionBoxes/index.tsx
@@ -1,0 +1,7 @@
+import { BridgingBox } from './BridgingBox'
+import { GeneralBox } from './GeneralBox'
+import { OtherQuestionBox } from './OtherQuestionBox'
+import { TrollFeeBox } from './TrollFeeBox'
+import { TrollQuestionBox } from './TrollQuestionBox'
+
+export { BridgingBox, GeneralBox, OtherQuestionBox, TrollFeeBox, TrollQuestionBox }
